// web/admin/src/types/api/models/CreateProducerRequest.ts
export interface CreateProducerRequest {
    /**
     * 姓(16文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    lastname: string;
    /**
     * 名(16文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    firstname: string;
    /**
     * 姓(かな)(ひらがな,32文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    lastnameKana: string;
    /**
     * 名(かな)(ひらがな,32文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    firstnameKana: string;
    /**
     * 店舗名(64文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    storeName: string;
    /**
     * サムネイルURL
     * @type {string}
     * @memberof CreateProducerRequest
     */
    thumbnailUrl?: string;
    /**
     * ヘッダー画像URL
     * @type {string}
     * @memberof CreateProducerRequest
     */
    headerUrl?: string;
    /**
     * メールアドレス
     * @type {string}
     * @memberof CreateProducerRequest
     */
    email: string;
    /**
     * 電話番号 (国際番号 + 電話番号)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    phoneNumber: string;
    /**
     * 郵便番号(ハイフンなし)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    postalCode: string;
    /**
     * 都道府県(32文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    prefecture: string;
    /**
     * 市区町村(32文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    city: string;
    /**
     * 町名・番地(64文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    addressLine1: string;
    /**
     * ビル名・号室など(64文字まで)
     * @type {string}
     * @memberof CreateProducerRequest
     */
    addressLine2: string;
}